+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/bookptm/DerivationBuilder.java
@@ -339,7 +339,7 @@ public class DerivationBuilder {
 
         builder.activity()
                 .id(defs.XID, "a7699").aka()
-                .label("unpacking activity")
+                .type(defs.Unpacking)
                 .build();
 
         builder.wasGeneratedBy()
@@ -698,6 +698,7 @@ public class DerivationBuilder {
                 .build();
 
         builder.specializationOf()
+                .id(defs.XID,"spe2").aka()
                 .generalEntity("b34")
                 .specificEntity("b34/6")
                 .attrQn(defs.specialization,"spe1")
@@ -849,7 +850,7 @@ public class DerivationBuilder {
                 .build();
 
         builder.entity()
-                .id(defs.XID, "item").aka()
+                .id(defs.XID, "item73").aka()
                 .label("An item")
                 .build();
 
@@ -869,14 +870,14 @@ public class DerivationBuilder {
         builder.used()
                 .activity("insert0")
                 .entity("coll0")
-                .role(PROVEXT, "inCollection")
+                .role(PROVEXT, "asCollection")
                 .attr(defs.dotColour, edge1Colour)
                 .build();
 
         builder.used()
                 .activity("insert0")
-                .entity("item")
-                .role(PROVEXT, "inEntity")
+                .entity("item73")
+                .role(PROVEXT, "asEntity")
                 .attr(defs.dotColour, edge1Colour)
                 .build();
 
@@ -899,7 +900,7 @@ public class DerivationBuilder {
 
         builder.hadMember()
                 .collection("coll1")
-                .entity("item")
+                .entity("item73")
                 .type(name.newProvQualifiedName("Derivation"))
                 .type(defs.InsertingElement)
                 .attr(defs.dotColour, edge3Colour)
@@ -922,13 +923,13 @@ public class DerivationBuilder {
                 .build();
 
         builder.entity()
-                .id(defs.XID, "item").aka()
+                .id(defs.XID, "item73").aka()
                 .label("An item")
                 .build();
 
         builder.hadMember()
                 .collection("coll0")
-                .entity("item")
+                .entity("item73")
                 .build();
 
         builder.entity()
@@ -947,14 +948,14 @@ public class DerivationBuilder {
         builder.used()
                 .activity("remove0")
                 .entity("coll0")
-                .role(PROVEXT, "inCollection")
+                .role(PROVEXT, "asCollection")
                 .attr(defs.dotColour, edge1Colour)
                 .build();
 
         builder.used()
                 .activity("remove0")
-                .entity("item")
-                .role(PROVEXT, "inEntity")
+                .entity("item73")
+                .role(PROVEXT, "asEntity")
                 .attr(defs.dotColour, edge1Colour)
                 .build();
 
@@ -969,7 +970,7 @@ public class DerivationBuilder {
 
         builder.wasDerivedFrom()
                 .generatedEntity("coll1")
-                .usedEntity("item")
+                .usedEntity("item73")
                 .activity("remove0")
                 .type(defs.RemovingElement)
                 .attr(defs.dotColour, edge3Colour)
@@ -1112,15 +1113,16 @@ public class DerivationBuilder {
         QualifiedName Transporting = builder.qn(VOCAB, "Transporting");
         QualifiedName Weighing     = builder.qn(VOCAB, "Weighing");
         QualifiedName Packing      = builder.qn(VOCAB, "Packing");
+        QualifiedName Unpacking      = builder.qn(VOCAB, "Unpacking");
         QualifiedName PickUp       = builder.qn(VOCAB, "PickUp");
         QualifiedName DropOff      = builder.qn(VOCAB, "DropOff");
         QualifiedName London       = builder.qn(VOCAB, "London");
         QualifiedName Brighton     = builder.qn(VOCAB, "Brighton");
         QualifiedName weight       = builder.qn(VOCAB, "weight");
         QualifiedName driver       = builder.qn(VOCAB, "Driver");
-        QualifiedName unpacker     = builder.qn(VOCAB, "unpacker");
-        QualifiedName scientist    = builder.qn(VOCAB, "scientist");
-        QualifiedName scale        = builder.qn(VOCAB, "scale");
+        QualifiedName unpacker     = builder.qn(VOCAB, "Unpacker");
+        QualifiedName scientist    = builder.qn(VOCAB, "Scientist");
+        QualifiedName scale        = builder.qn(VOCAB, "Scale");
         QualifiedName Instrument   = builder.qn(VOCAB, "Instrument");
         QualifiedName specialization=builder.qn(PROVEXT, "specialization");
         QualifiedName derivation    =builder.qn(PROVEXT, "derivation");
@@ -1138,7 +1140,7 @@ public class DerivationBuilder {
 
 
         QualifiedName foaf_name    = builder.qn(FOAF, "name");
-        Definitions result = new Definitions(XID, Box, FPC, Ownership, Recycling, Transporting, Weighing, Packing, PickUp, DropOff, London, Brighton, weight, foaf_name, driver, unpacker, scientist, scale, Instrument, specialization, derivation, entity, association, generation, activity, invalidation, usage, dotColour, InsertingElement, InsertingIntoCollection, RemovingElement, RemovingElementFromCollection);
+        Definitions result = new Definitions(XID, Box, FPC, Ownership, Recycling, Transporting, Weighing, Packing, Unpacking, PickUp, DropOff, London, Brighton, weight, foaf_name, driver, unpacker, scientist, scale, Instrument, specialization, derivation, entity, association, generation, activity, invalidation, usage, dotColour, InsertingElement, InsertingIntoCollection, RemovingElement, RemovingElementFromCollection);
         return result;
     }
 
@@ -1175,8 +1177,9 @@ public class DerivationBuilder {
         public final QualifiedName InsertingIntoCollection;
         public final QualifiedName RemovingElement;
         public final QualifiedName RemovingElementFromCollection;
+        private final QualifiedName Unpacking;
 
-        public Definitions(Prefix XID, QualifiedName Box, QualifiedName fpc, QualifiedName Ownership, QualifiedName Recycling, QualifiedName Transporting, QualifiedName Weighing, QualifiedName packing, QualifiedName PickUp, QualifiedName DropOff, QualifiedName London, QualifiedName Brighton, QualifiedName weight, QualifiedName foaf_name, QualifiedName driver, QualifiedName unpacker, QualifiedName scientist, QualifiedName scale, QualifiedName instrument, QualifiedName specialization, QualifiedName derivation, QualifiedName entity, QualifiedName association, QualifiedName generation, QualifiedName activity, QualifiedName invalidation, QualifiedName usage, QualifiedName dotColour, QualifiedName insertingElement, QualifiedName insertingIntoCollection, QualifiedName removingElement, QualifiedName removingElementFromCollection) {
+        public Definitions(Prefix XID, QualifiedName Box, QualifiedName fpc, QualifiedName Ownership, QualifiedName Recycling, QualifiedName Transporting, QualifiedName Weighing, QualifiedName packing, QualifiedName unpacking, QualifiedName PickUp, QualifiedName DropOff, QualifiedName London, QualifiedName Brighton, QualifiedName weight, QualifiedName foaf_name, QualifiedName driver, QualifiedName unpacker, QualifiedName scientist, QualifiedName scale, QualifiedName instrument, QualifiedName specialization, QualifiedName derivation, QualifiedName entity, QualifiedName association, QualifiedName generation, QualifiedName activity, QualifiedName invalidation, QualifiedName usage, QualifiedName dotColour, QualifiedName insertingElement, QualifiedName insertingIntoCollection, QualifiedName removingElement, QualifiedName removingElementFromCollection) {
             this.XID = XID;
             this.Box = Box;
             this.FPC = fpc;
@@ -1185,6 +1188,7 @@ public class DerivationBuilder {
             this.Transporting = Transporting;
             this.Weighing = Weighing;
             this.Packing = packing;
+            this.Unpacking = unpacking;
             this.PickUp = PickUp;
             this.DropOff = DropOff;
             this.London = London;