+++ RELEASES.md
@@ -1,6 +1,9 @@
 See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.2.2-SNAPSHOT`
+* prov-template: added ptm-merging template
+* prov-template-library: added ptm-merging template
+* prov-template: updated BatchExecutor to create provenance when merging templates
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java